[bidding_service] - controllers refactoring
