+++ src/main/java/com/kirill/auction/bot/dto/BidDto.java
@@ -1,4 +1,4 @@
-package com.kirill.auction.bot.model;
+package com.kirill.auction.bot.dto;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
@@ -6,7 +6,7 @@ import io.swagger.annotations.ApiModelProperty;
 import java.util.Objects;
 
 @ApiModel("Contains the bids of the two bidders.")
-public class BidsRequest {
+public class BidDto {
 
     @ApiModelProperty("The bid of this bidder")
     private int own;
@@ -14,7 +14,7 @@ public class BidsRequest {
     @ApiModelProperty("The bid of the other bidder")
     private int other;
 
-    public BidsRequest(final int own, final int other) {
+    public BidDto(final int own, final int other) {
         this.own = own;
         this.other = other;
     }
@@ -30,8 +30,8 @@ public class BidsRequest {
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (!(o instanceof BidsRequest)) return false;
-        BidsRequest that = (BidsRequest) o;
+        if (!(o instanceof BidDto)) return false;
+        BidDto that = (BidDto) o;
         return own == that.own &&
                 other == that.other;
     }
